+++ src/interactions/col_nil.c
@@ -22,7 +22,7 @@
 Term col_nil(Term col, Term nil) {
   interaction_count++;
   //printf("col_nil\n");
-  uint32_t col_loc = TERM_VAL(col);
+  uint64_t col_loc = TERM_VAL(col);
 
   // Create the nil term to be returned and for substitution
   Term nil_term = make_term(NIL, 0, 0);
