+++ src/interactions/use_sup.c
@@ -21,8 +21,8 @@
 Term use_sup(Term use, Term sup) {
   interaction_count++;
   //printf("use_sup\n");
-  uint32_t use_loc = TERM_VAL(use);
-  uint32_t sup_loc = TERM_VAL(sup);
+  uint64_t use_loc = TERM_VAL(use);
+  uint64_t sup_loc = TERM_VAL(sup);
   uint8_t  sup_lab = TERM_LAB(sup);
 
   Term a = heap[sup_loc + 0];
@@ -30,20 +30,20 @@ Term use_sup(Term use, Term sup) {
   Term k = heap[use_loc + 1];
 
   // Create a collapser for k
-  uint32_t col_loc = alloc(1);
+  uint64_t col_loc = alloc(1);
   heap[col_loc] = k;
 
   // Create new USE nodes for each branch
-  uint32_t use_a_loc = alloc(2);
+  uint64_t use_a_loc = alloc(2);
   heap[use_a_loc + 0] = a;
   heap[use_a_loc + 1] = make_term(CO0, sup_lab, col_loc);
 
-  uint32_t use_b_loc = alloc(2);
+  uint64_t use_b_loc = alloc(2);
   heap[use_b_loc + 0] = b;
   heap[use_b_loc + 1] = make_term(CO1, sup_lab, col_loc);
 
   // Create new superposition with the USE nodes
-  uint32_t new_sup_loc = alloc(2);
+  uint64_t new_sup_loc = alloc(2);
   heap[new_sup_loc + 0] = make_term(USE, 0, use_a_loc);
   heap[new_sup_loc + 1] = make_term(USE, 0, use_b_loc);
