+++ src/parse/term/bit.c
@@ -3,7 +3,7 @@
 #include "../../memory.h"
 
 // Parse a bool type
-void parse_term_bit(Parser* parser, uint32_t loc) {
+void parse_term_bit(Parser* parser, uint64_t loc) {
   if (check_utf8_4bytes(parser, 0xF0, 0x9D, 0x94, 0xB9)) {
     consume_utf8(parser, 4);
   } else if (!consume(parser, "𝔹")) {
