+++ src/parse/term/ite.c
@@ -3,13 +3,13 @@
 #include "../../memory.h"
 
 // Parse a bool elimination
-void parse_term_ite(Parser* parser, uint32_t loc) {
+void parse_term_ite(Parser* parser, uint64_t loc) {
   expect(parser, "?", "for bool elimination");
 
-  uint32_t ite_node = alloc_term(3);
-  uint32_t cnd_loc = ite_node;
-  uint32_t thn_loc = ite_node + 1;
-  uint32_t els_loc = ite_node + 2;
+  uint64_t ite_node = alloc_term(3);
+  uint64_t cnd_loc = ite_node;
+  uint64_t thn_loc = ite_node + 1;
+  uint64_t els_loc = ite_node + 2;
 
   parse_term(parser, cnd_loc);
   expect(parser, "{", "after condition in bool elimination");
