+++ src/parse/term/sig.c
@@ -3,7 +3,7 @@
 #include "../../memory.h"
 
 // Parse a sigma type
-void parse_term_sig(Parser* parser, uint32_t loc) {
+void parse_term_sig(Parser* parser, uint64_t loc) {
   if (check_utf8(parser, 0xCE, 0xA3)) {
     consume_utf8(parser, 2);
   } else if (!consume(parser, "Σ")) {
@@ -13,9 +13,9 @@ void parse_term_sig(Parser* parser, uint32_t loc) {
   char* name = parse_name(parser);
   expect(parser, ":", "after name in sigma type");
 
-  uint32_t sig_node = alloc_term(2);
-  uint32_t fst_loc = sig_node;
-  uint32_t snd_loc = sig_node + 1;
+  uint64_t sig_node = alloc_term(2);
+  uint64_t fst_loc = sig_node;
+  uint64_t snd_loc = sig_node + 1;
 
   parse_term(parser, fst_loc);
   expect(parser, ".", "after first type in sigma type");
